Reestructure backend
